requirements.txt
flask==1.1.2
flask-restful
flask-sqlalchemy
flask-migrate
flask-swagger-ui
marshmallow-sqlalchemy
pyjwt